requirements.txt
# requirements.txt
requests
beautifulsoup4
streamlit
openai
python-dotenv
pandas
anthropic
numpy
matplotlib
